Fixed usage of shared gauges, counters and histograms. Previously gauges, counters and histograms shared their name but were registered more than once. The register calls failed in prometheus' simpleclient but these errors were silently ignored. The result was that only the first collector was used for a name. Metrics like HTTPServerPrometheusMetrics are instantiated multiple times and thus need to share gauges, counters and histograms to work properly across different instances of HTTPServers, HTTPClients etc.
This commit adds factories for shared gauges, counters and histograms.
These are created in VertxPrometheusMetrics and passed around to the different child metrics.
The factories are closed in VertxPrometheusMetrics::close(). They can not
be closed earlier as the collectors need to be registered until the last
client (HTTPServerMetrics, ...) is released. The only safe point to to this
is the initial VertxPrometheusMetrics.